+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [1.0.8]
 
 - Updated dependencies - `jetleaf_build`
+- Added `getAsset` method in `StringExtension`
 
 ## [1.0.7]